+++ lib/skeem/primitive/primitive_builder.rb
@@ -1,19 +1,26 @@
-require_relative '../primitive_func'
+require_relative '../primitive_procedure'
 
 module Skeem
   module Primitive
     module PrimitiveBuilder
       def add_primitives(aRuntime)
         add_arithmetic(aRuntime)
+        add_number_predicates(aRuntime)
       end
       
       private
       
       def add_arithmetic(aRuntime)
-        def_func(aRuntime, create_plus)
-        def_func(aRuntime, create_minus)
-        def_func(aRuntime, create_multiply)
-        def_func(aRuntime, create_divide)
+        def_procedure(aRuntime, create_plus)
+        def_procedure(aRuntime, create_minus)
+        def_procedure(aRuntime, create_multiply)
+        def_procedure(aRuntime, create_divide)
+      end
+      
+      def add_number_predicates(aRuntime)
+        def_procedure(aRuntime, create_number?)
+        def_procedure(aRuntime, create_real?)
+        def_procedure(aRuntime, create_integer?)
       end
       
       def create_plus()
@@ -22,11 +29,7 @@ module Skeem
           operands = arglist.tail.to_eval_enum(runtime)
           raw_result = first_one.value
           operands.each { |elem| raw_result += elem.value }
-          if raw_result.kind_of?(Float) 
-            SExprReal.create(raw_result)
-          else
-            SExprInteger.create(raw_result)
-          end
+          to_skm(raw_result)
         end
 
         ['+', plus_code]
@@ -38,11 +41,7 @@ module Skeem
           operands = arglist.tail.to_eval_enum(runtime)
           raw_result = first_one.value
           operands.each { |elem| raw_result -= elem.value } 
-          if raw_result.kind_of?(Float) 
-            SExprReal.create(raw_result)
-          else
-            SExprInteger.create(raw_result)
-          end
+          to_skm(raw_result)
         end
 
         ['-', minus_code]
@@ -54,11 +53,7 @@ module Skeem
           operands = arglist.tail.to_eval_enum(runtime)
           raw_result = first_one.value
           operands.each { |elem| raw_result *= elem.value }          
-          if raw_result.kind_of?(Float) 
-            SExprReal.create(raw_result)
-          else
-            SExprInteger.create(raw_result)
-          end
+          to_skm(raw_result)
         end
 
         ['*', multiply_code]
@@ -70,23 +65,58 @@ module Skeem
           operands = arglist.tail.to_eval_enum(runtime)
           raw_result = first_one.value
           operands.each { |elem| raw_result /= elem.value }
-          if raw_result.kind_of?(Float) 
-            SExprReal.create(raw_result)
-          else
-            SExprInteger.create(raw_result)
-          end
+          to_skm(raw_result)
         end
 
         ['/', divide_code]
+      end
+
+      def create_number?()
+        pred_code = ->(runtime, arg) do
+          arg_evaluated = arg.evaluate(runtime)
+          to_skm(arg_evaluated.number?)
+        end
+
+        ['number?', pred_code]
+      end
+
+      def create_real?()
+        pred_code = ->(runtime, arg) do
+          arg_evaluated = arg.evaluate(runtime)
+          to_skm(arg_evaluated.real?)
+        end
+
+        ['real?', pred_code]
+      end
+
+      def create_integer?()
+        pred_code = ->(runtime, arg) do
+          arg_evaluated = arg.evaluate(runtime)
+          to_skm(arg_evaluated.integer?)
+        end
+
+        ['integer?', pred_code]
       end       
       
-      def def_func(aRuntime, aPair)
-        func = PrimitiveFunc.new(aPair.first, aPair.last)
+      def def_procedure(aRuntime, aPair)
+        func = PrimitiveProcedure.new(aPair.first, aPair.last)
         define(aRuntime, func.identifier, func)
       end
 
       def define(aRuntime, aKey, anEntry)
         aRuntime.define(aKey, anEntry)
+      end
+      
+      # Convert Ruby object into its Skeem counterpart
+      def to_skm(native_obj)
+        case native_obj
+          when TrueClass, FalseClass
+            SkmBoolean.create(native_obj)
+          when Float 
+            SkmReal.create(native_obj)
+          when Integer
+            SkmInteger.create(native_obj)     
+        end
       end
     end # module
   end # module